Add delete function for Duplicates
Every time onClickMyRep runs, deleteDuplicatesForUserTunes runs which:
stores the Index data as store.tunes, then sorts them alphabetically,
checking if there are duplicates. If there is a duplicate, api.delete
deletes that duplicate and onClickMyRep is run again until there are
no duplicates. If user creates a duplicate, it doesn't get deleted. It
simply doesn't show because of ui line 83. This does not effect the
shared tune function.